requirements.txt
uvicorn~=0.23.2
fastapi~=0.104.0
pydantic_settings~=2.0.3
requests~=2.31.0
apscheduler~=3.10.0
numpy~=1.26.1
scikit-learn~=1.3.1
astropy~=5.3.4
Pillow~=10.1.0
opencv-python~=4.8.1.78
matplotlib~=3.8.0
scipy~=1.11.3